refactor: moved migration validator to std & added comments to the crowdfund query
